requirements.txt
requests==2.20.0
feedparser
yara-python
newspaper3k
